@@ -1,15 +1,15 @@
 package br.uff.tecnomarias.rest.dto;
 
-
 import br.uff.tecnomarias.domain.entity.Links;
 import br.uff.tecnomarias.domain.entity.PessoaFisica;
+import br.uff.tecnomarias.domain.entity.Vaga;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class PessoaFisicaDTO extends PessoaDTO {
 
-    private Links links;
+    public Links links;
 
     public PessoaFisicaDTO() {
         super();
@@ -22,22 +22,12 @@ public class PessoaFisicaDTO extends PessoaDTO {
 
     public PessoaFisica toEntity() {
         PessoaFisica pf = new PessoaFisica();
-        if(this.links != null)
-            pf.setLinks(this.links);
+        pf.setLinks(this.links);
         gerarPessoa(pf);
         return pf;
     }
 
     public static List<PessoaFisicaDTO> toDTOList(List<PessoaFisica> pfList) {
-        return pfList.stream().map(pf -> new PessoaFisicaDTO(pf)).collect(Collectors.toList());
+        return pfList.stream().map(PessoaFisicaDTO::new).collect(Collectors.toList());
     }
-
-    public Links getLinks() {
-        return links;
-    }
-
-    public void setLinks(Links links) {
-        this.links = links;
-    }
-
 }
