@@ -1,15 +1,18 @@
 package br.uff.tecnomarias.rest.dto;
 
 import br.uff.tecnomarias.domain.entity.Feedback;
+import br.uff.tecnomarias.domain.entity.Vaga;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class FeedbackDTO {
 
-    private Long id;
-    private String comentario;
-    private Long idPessoa;
+    public Long id;
+    public String comentario;
+    public Long idPessoa;
 
-    public FeedbackDTO() {
-    }
+    public FeedbackDTO() { }
 
     public FeedbackDTO(Feedback feedback) {
         this.id = feedback.getId();
@@ -22,28 +25,7 @@ public class FeedbackDTO {
         return feedback;
     }
 
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getComentario() {
-        return comentario;
-    }
-
-    public void setComentario(String comentario) {
-        this.comentario = comentario;
-    }
-
-    public Long getIdPessoa() {
-        return idPessoa;
-    }
-
-    public void setIdPessoa(Long idPessoa) {
-        this.idPessoa = idPessoa;
+    public static List<FeedbackDTO> toDTOList(List<Feedback> vagas) {
+        return vagas.stream().map(FeedbackDTO::new).collect(Collectors.toList());
     }
 }
-
