@@ -1,43 +1,54 @@
 package br.uff.tecnomarias.rest.resource;
 
-
+import br.uff.tecnomarias.domain.entity.PessoaFisica;
 import br.uff.tecnomarias.rest.dto.PessoaFisicaDTO;
-import br.uff.tecnomarias.rest.dto.PessoaJuridicaDTO;
 import br.uff.tecnomarias.service.PessoaFisicaService;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
 
-import javax.inject.Inject;
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
+import java.util.List;
 
-@Path("pf")
-@Produces(MediaType.APPLICATION_JSON)
-@Consumes(MediaType.APPLICATION_JSON)
+@RestController
+@RequestMapping("pf")
+@Tag(name = "Pessoa Fisica")
+@CrossOrigin
 public class PessoaFisicaResource {
 
-    @Inject
+    @Autowired
     PessoaFisicaService pfService;
 
-    @POST
-    public Response salvarPF(PessoaFisicaDTO pfDTO) {
-        PessoaFisicaDTO pfSalva = new PessoaFisicaDTO(pfService.salvar(pfDTO.toEntity()));
-        return Response.ok(pfSalva).build();
+    @PostMapping
+    @ResponseBody
+    public PessoaFisicaDTO criarPF(@RequestBody PessoaFisicaDTO pf) {
+        return new PessoaFisicaDTO(pfService.salvar(pf.toEntity()));
     }
 
-    @PUT
-    @Path("{id}")
-    public Response alterar(@PathParam("id") final Long id, PessoaFisicaDTO pfDTO) {
-        return Response.ok(new PessoaFisicaDTO(pfService.alterar(id, pfDTO.toEntity()))).build();
+    @GetMapping
+    @ResponseBody
+    public List<PessoaFisicaDTO> buscarTodas() {
+        List<PessoaFisica> pfList = pfService.buscarTodas();
+        return PessoaFisicaDTO.toDTOList(pfList);
     }
 
-    @GET
-    public Response buscarTodas() {
-        return Response.ok(PessoaFisicaDTO.toDTOList(pfService.buscarTodas())).build();
+    @GetMapping("{id}")
+    @ResponseBody
+    public PessoaFisicaDTO buscarPorId(@PathVariable Long id) {
+        return new PessoaFisicaDTO(pfService.buscarPorId(id));
     }
 
-    @GET
-    @Path("{id}")
-    public Response buscarPorId(@PathParam("id") final Long id) {
-        return Response.ok(new PessoaFisicaDTO(pfService.buscarPorId(id))).build();
+    @PutMapping("{id}")
+    @ResponseBody
+    public PessoaFisicaDTO alterar(@PathVariable Long id,
+                                   @RequestBody PessoaFisicaDTO pfDTO) {
+        return new PessoaFisicaDTO(pfService.alterar(id, pfDTO.toEntity()));
     }
+
+    @DeleteMapping("{id}")
+    @ResponseBody
+    public String remover(@PathVariable Long id) {
+        pfService.remover(id);
+        return "Pessoa removida com sucesso.";
+    }
+
 }
