@@ -3,67 +3,62 @@ package br.uff.tecnomarias.rest.resource;
 import br.uff.tecnomarias.domain.entity.Vaga;
 import br.uff.tecnomarias.rest.dto.VagaDTO;
 import br.uff.tecnomarias.service.VagaService;
+import br.uff.tecnomarias.service.exception.BadRequestException;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.HttpClientErrorException;
 
-import javax.inject.Inject;
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-@Path("vaga")
-@Produces(MediaType.APPLICATION_JSON)
-@Consumes(MediaType.APPLICATION_JSON)
+@RestController
+@RequestMapping("vaga")
+@Tag(name = "Vaga")
+@CrossOrigin
 public class VagaResource {
 
-    @Inject
+    @Autowired
     VagaService vagaService;
 
-    @POST
-    public Response criarVaga(VagaDTO vagaDTO) {
-        VagaDTO vagaSalva = new VagaDTO(
-                vagaService.salvar(
-                        vagaDTO.toEntity()
-                )
-        );
-        return Response.status(Response.Status.CREATED).entity(vagaSalva).build();
+    @PostMapping
+    @ResponseBody
+    public VagaDTO criarVaga(@RequestBody VagaDTO vagaDTO) {
+        return new VagaDTO(vagaService.salvar(vagaDTO.toEntity()));
     }
 
-    @GET
-    public Response buscarTodas(@QueryParam("start") final int start,
-                                @QueryParam("pageSize") final int size) {
-        List<Vaga> vagas = vagaService.buscarTodas(start, size);
-        return Response.ok(vagas.stream().map(vaga -> new VagaDTO(vaga)).collect(Collectors.toList())).build();
+    @GetMapping
+    @ResponseBody
+    public List<VagaDTO> buscarTodas() {
+        List<Vaga> vagas = vagaService.buscarTodas();//start, pageSize);
+        return VagaDTO.toDTOList(vagas);
     }
 
-    @GET
-    @Path("{id}")
-    public Response buscarPorId(@PathParam("id") final Long id) {
-        Vaga vaga = vagaService.buscarPorId(id);
-        Response.Status status = vaga != null ? Response.Status.OK : Response.Status.NO_CONTENT;
-        return Response.status(status).entity(new VagaDTO(vaga)).build();
+    @GetMapping("{id}")
+    @ResponseBody
+    public VagaDTO buscarPorId(@PathVariable Long id) {
+        return new VagaDTO(vagaService.buscarPorId(id));
     }
 
-    @GET
-    @Path("count")
-    @Produces(MediaType.TEXT_PLAIN)
-    public Response buscarTotal() {
-        return Response.ok(vagaService.count()).build();
+    @GetMapping("count")
+    @ResponseBody
+    public Long buscarTotal() {
+        return vagaService.count();
     }
 
-    @GET
-    @Path("empresa/{id}")
-    public Response buscarPorEmpresa(@PathParam("id") final Long id) {
+    @GetMapping("empresa/{id}")
+    @ResponseBody
+    public List<VagaDTO> buscarPorEmpresa(@PathVariable Long id) {
         List<Vaga> vagas = vagaService.buscarPorEmpresa(id);
-        return Response.ok(VagaDTO.toDTOList(vagas)).build();
+        return VagaDTO.toDTOList(vagas);
     }
 
-    @GET
-    @Path("filtro")
-    public Response buscarPorFiltro(@QueryParam("filtro") final String filtro,
-                                    @QueryParam("valor") final String valor) {
+    @GetMapping("filtro")
+    @ResponseBody
+    public List<VagaDTO> buscarPorFiltro(@RequestParam String filtro,
+                                         @RequestParam final String valor) {
         List<Vaga> vagas;
         switch (filtro) {
             case "areaAtuacao":
@@ -76,32 +71,30 @@ public class VagaResource {
                 vagas = vagaService.buscarPorLocalidade(valor);
                 break;
             default:
-                return Response.status(Response.Status.BAD_REQUEST).build();
+                throw new BadRequestException("Filtro invalido");
         }
-        return Response.ok(VagaDTO.toDTOList(vagas)).build();
+        return VagaDTO.toDTOList(vagas);
     }
 
-    @PUT
-    @Path("{id}")
-    public Response alterar(@PathParam("id") final Long id,
-                            final VagaDTO vagaDTO) {
-        return Response.ok(new VagaDTO(
-                vagaService.alterar(id, vagaDTO.toEntity())
-        )).build();
+    @PutMapping("{id}")
+    @ResponseBody
+    public VagaDTO alterar(@PathVariable Long id,
+                           @RequestBody VagaDTO vagaDTO) {
+        return new VagaDTO(vagaService.alterar(id, vagaDTO.toEntity()));
     }
 
-    @DELETE
-    @Path("{id}")
-    public Response removerVaga(@PathParam("id") final Long id) {
+    @DeleteMapping("{id}")
+    @ResponseBody
+    public String removerVaga(@PathVariable Long id) {
         vagaService.remover(id);
-        return Response.ok("Vaga removida com sucesso.").build();
+        return "Vaga removida com sucesso.";
     }
 
-    @GET
-    @Path("area_atuacao")
-    public Response listarAreaAtuacao() {
+    @GetMapping("area_atuacao")
+    @ResponseBody
+    public Map<String, List<String>> listarAreaAtuacao() {
         Map<String, List<String>> resp = new HashMap<>();
         resp.put("listaAreaAtuacao", vagaService.listarAreaAtuacao());
-        return Response.ok(resp).build();
+        return resp;
     }
 }
